+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+- Add support for `Elo` card type
 - Adds `updateCard` method (#77)
 
 7.0.0